moved RangerHdfsAccessRequest, RangerHdfsResource
